feat: 무한 스크롤를 컨트롤러에서 분리
